Created abstract ancestor for explorer items (fixes #838)
